pyproject.toml
[project]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksii Ablitsov",email = "oleksii.ablitsov@bringg.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<1.8.0)",
    "bcrypt (>=4.0.1,<4.1.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.43.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sphinx (>=8.2.3,<9.0.0)",
    "sphinx-rtd-theme (>=3.0.2,<4.0.0)",
    "sphinx-autodoc-typehints (>=3.1.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
